Reorganized module part of directory
