+++ engine/CMakeLists.txt
@@ -2,6 +2,6 @@
         BlockyEngine.cpp
 )
 
-add_subdirectory(modules)
+add_subdirectory(moduleManager)
 
 target_include_directories(${PROJECT_NAME} PUBLIC ${CMAKE_CURRENT_SOURCE_DIR})
