+++ engine/moduleManager/CMakeLists.txt
@@ -2,3 +2,5 @@
         ModuleManager.cpp
         ModuleWrapper.cpp
 )
+
+add_subdirectory(modules)
