+++ engine/moduleManager/modules/CMakeLists.txt
@@ -1,5 +1,3 @@
 ﻿target_sources(${PROJECT_NAME} PRIVATE
         WindowModule.cpp
 )
-
-add_subdirectory(core)
